pyproject.toml
[project]
name = "simulative-ml-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.49.1",
    "gspread>=6.2.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
]
